componentization and styled components
