--- App.css
@@ -1,72 +1,0 @@
-.container {
-  max-width: 600px;
-  margin: 0 auto;
-  padding: 20px;
-  font-family: Arial, sans-serif;
-  text-align: center;
-}
-
-.title {
-  color: #333;
-  margin-bottom: 20px;
-}
-
-.filterContainer {
-  display: flex;
-  gap: 20px;
-  justify-content: center;
-  margin-bottom: 20px;
-  flex-direction: row;
-}
-
-.filterItem {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-
-.label {
-  margin-bottom: 5px;
-  font-weight: bold;
-}
-
-.input {
-  padding: 8px;
-  width: 200px;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-}
-
-.button {
-  padding: 10px 20px;
-  margin-top: 10px;
-  background-color: #003d6b;
-  color: #fff;
-  border: none;
-  border-radius: 4px;
-}
-
-.list {
-  list-style-type: none;
-  padding: 0;
-}
-
-.listItem {
-  background-color: #f9f9f9;
-  padding: 10px;
-  margin: 10px 0;
-  border-radius: 4px;
-  box-shadow: 0 2px 5px rgba(0, 0, 0, 0.1);
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.userName {
-  font-weight: bold;
-}
-
-.userEmail {
-  margin-left: 10px;
-  color: #777;
-}
