+++ src/App.css
@@ -1,4 +1,4 @@
-.container {
+/* .container {
   max-width: 600px;
   margin: 0 auto;
   padding: 20px;
@@ -69,4 +69,4 @@
 .userEmail {
   margin-left: 10px;
   color: #777;
-}
+} */
